az-ml/titanic/requirements.txt
azureml-sdk
python-dotenv
